netdev feedback from David Miller:
A pointer is not castable directly to a u32 everywhere.  
It can be 64-bits and in which case GCC will warn here.
Use "unsigned long" or similar.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@6496 526b6c2d-f592-4532-a319-5dd88ccb003d

+++ drivers/net/bfin_mac.c
@@ -617,7 +617,7 @@ static int bfin_mac_hard_start_xmit(struct sk_buff *skb,
 				struct net_device *dev)
 {
 	u16 *data;
-	u32 data_align = (u32)(skb->data) & 0x3;
+	u32 data_align = (unsigned long)(skb->data) & 0x3;
 	current_tx_ptr->skb = skb;
 
 	if (data_align == 0x2) {
